Move RefinanceTokenModal and RepayTokenModal to TokenLoansContent
